// src/common/types/apiResponse.d.ts
type ApiResponseID =
  | '00'
  | '01'
  | '02'
  | '03'
  | '04'
  | '05'
  | '06'
  | '07'
  | '08'
  | '09'
  | '10'
  | '11'
  | '12'
  | '13'
  | '14'
  | '15';

export interface ApiResponse<Data> {
  id: ApiResponseID;
  msg: string;
  data?: Data;
}

// Generated by https://quicktype.io

export interface APIResponsePaginated<Data> {
  id: string;
  msg: string;
  data: DataPaginated<Data>;
}

export interface DataPaginated<Data> {
  jumlah: string;
  data?: Data[];
}